requirements.txt
boto3
azure-storage-blob
flask
fastapi==0.109.0
uvicorn==0.27.0
email-validator==2.1.0.post1
dkimpy==1.0.5
pyspf==2.0.14
requests==2.31.0
redis==5.0.1
python-dotenv==1.0.1
python-whois==0.8.0
python-magic==0.4.27
argparse==1.4.0